feat: Implement customer authentication, dedicated customer order and reservation management, an admin dashboard, and comprehensive project documentation.

+++ src/main/java/edu/qsp/restorent_management_system/repository/MenuRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import edu.qsp.restorent_management_system.model.MenuItem;
 
-public interface  MenuRepository extends JpaRepository<MenuItem , Integer>{
-    
+public interface MenuRepository extends JpaRepository<MenuItem, Long> {
+
 }
